Cleared a bunch of warnings, added small fix to make target bitrate estimation more robust to media coder inactivity

+++ code/ScreamTx.h
@@ -156,7 +156,7 @@ public:
 	/*
 	* Print logs
 	*/
-	void printLog(double time);
+	void printLog(float time);
 
 private:
 	/*
